회전베기 스킬 구현

+++ Assets/Scripts/Item/Weapon/HitDetection.cs
@@ -8,7 +8,7 @@ public class HitDetection : MonoBehaviour
     //[field: SerializeField] public WeaponAttribute addedWeaponAttribute { get; private set; } //부가 속성?
     [field: SerializeField] public float damage { get; private set; }
     [field: SerializeField] public float knockBack { get; private set; }
-    [field: SerializeField] public int critical { get; private set; }
+    [field: SerializeField] public float critical { get; private set; }
     [field: SerializeField] public float criticalDamage { get; private set; }
     //[field: SerializeField] public int drain { get; private set; }
 
@@ -20,7 +20,7 @@ public class HitDetection : MonoBehaviour
     /// <param name="knockBack"></param>
     /// <param name="critical"></param>
     /// <param name="criticalDamage"></param>
-    public void SetHitDetection(WeaponAttribute weaponAttribute, float damage, float knockBack, int critical, float criticalDamage)
+    public void SetHitDetection(WeaponAttribute weaponAttribute, float damage, float knockBack, float critical, float criticalDamage)
     {
         this.damage = damage;
         this.weaponAttribute = weaponAttribute;
